requirements.txt
aenum==3.1.15
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Deprecated==1.2.14
distro==1.9.0
fastapi==0.110.0
frozenlist==1.4.1
future==1.0.0
google-api-core==2.18.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-cloud-vision==3.7.2
googleapis-common-protos==1.63.0
grpcio==1.62.1
grpcio-status==1.62.1
gspread==6.0.2
h11==0.14.0
httpcore==1.0.4
httplib2==0.22.0
httpx==0.27.0
idna==3.6
line-bot-sdk==3.9.0
multidict==6.0.5
oauth2client==4.1.3
oauthlib==3.2.2
openai==1.14.2
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.6.4
pydantic_core==2.16.3
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
starlette==0.36.3
StrEnum==0.4.15
tqdm==4.66.2
typing_extensions==4.10.0
urllib3==2.2.1
uvicorn==0.29.0
wrapt==1.16.0
yarl==1.9.4
